Rename MissingReferences to MissingProjectReferences; Introduce MissingReferencesWindow base class for shared code between MissingProjectReferences and MissingSceneReferences; Add comments to explain what is going on in MissingReferencesWindow
