+++ Missingreferences/Editor/MissingReferencesWindow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 469fc51729da5f64799675dc789d9343
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
